backup/requirements.txt
functions-framework==3.*
google-api-python-client
google-auth
google-cloud-compute
Flask
markupsafe
pytz
google-api-core